+++ src/troupes/Chevalier.java
@@ -6,7 +6,7 @@ import royaume.Constantes;
 
 public class Chevalier extends Troupe {
 	
-	public Chevalier(Pane layer, Image image, double pos_x, double pos_y) {
+	public Chevalier(Pane layer, Image image, long pos_x, long pos_y) {
 		super(layer, image, 500, 20, 6, Constantes.VIE_CHEVALIER, 5, pos_x, pos_y);
 	}
