+++ src/troupes/Piquier.java
@@ -6,7 +6,7 @@ import royaume.Constantes;
 
 public class Piquier extends Troupe {
 
-	public Piquier(Pane layer, Image image, double pos_x, double pos_y) {
+	public Piquier(Pane layer, Image image, long pos_x, long pos_y) {
 		super(layer, image, 100, 5, 2, Constantes.VIE_PIQUIER, 1, pos_x, pos_y);
 	}
